Create the Scrambler component
This is how I'm calling both the Plugboard and the Rotor, because their
purpose is to scramble the letters from the message in the machine.
The Reflector just bounces the signal back, mirroring the letter
pairings it receives.

It's obvious that the project is still not done, since the encryption
isn't implemented yet. This is a savepoint where all the implemented
tests are passing. We will move to the next test in the next iteration.