Implement Reflector component
A pattern is starting to emerge here. The Reflector doesn't seem so
different from the other components. Except for the wiring, which is
actually the base of the system.

Another thing, I think we could create a Wiring object later, to make it
easier to initialize and rotate the tables.